requirements.txt
faiss-cpu
torch
transformers
flask
PyPDF2
protobuf